[FE-1235-2] Moved ActiveFilters to src/components/reportBuilder
