+++ src/components/reportBuilder/index.js
@@ -1,3 +1,4 @@
+export { default as ActiveFilters } from './ActiveFilters';
 export { default as AggregatedMetrics } from './AggregatedMetrics';
 export { default as CompareByAggregatedMetrics } from './CompareByAggregatedMetrics';
 export { ChartGroups, Charts } from './Charts';
