+++ src/pages/reportBuilder/components/ReportOptions.js
@@ -2,12 +2,12 @@ import React, { useEffect, useState, useCallback, useMemo } from 'react';
 import { connect } from 'react-redux';
 import { Box, Button, Drawer, Expandable, Inline, Panel, Stack } from 'src/components/matchbox';
 import { Tabs, Loading } from 'src/components';
+import { ActiveFilters } from 'src/components/reportBuilder';
 import { useReportBuilderContext } from '../context/ReportBuilderContext';
 import { selectFeatureFlaggedMetrics } from 'src/selectors/metrics';
 import { parseSearchNew as parseSearch } from 'src/helpers/reports';
 import { isAccountUiOptionSet } from 'src/helpers/conditions/account';
 import {
-  ActiveFilters,
   ActiveMetrics,
   ActiveComparisons,
   CompareByForm,
